TAKE3/requirements.txt
# Core RL + environment
gymnasium>=0.29.1
torch>=2.1.0
numpy>=1.24.0

# Visualisation & analysis
matplotlib>=3.7.0

# GIF generation (visualize.py)
pillow>=9.0.0
imageio>=2.28.0
pygame>=2.4.0     # required for rgb_array rendering in gymnasium classic control
